requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-openai>=0.0.5
langchain-astradb>=0.1.0
langchain-core>=0.1.7
astrapy>=0.5.2 